Adds missing XMSS signature length check.
- Fixes out of bounds read in `XMSS_Signature` constructor when the raw
  signature data supplied as arguments is shorter than the signature size
  defined by the XMSS parameter set encoded in the `XMSS_PublicKey`.
- Fixes valid signatures with arbitrary appended data to be verified as
  correct signature.

+++ src/lib/pubkey/xmss/xmss_signature.h
@@ -117,7 +117,7 @@ class XMSS_Signature final
       secure_vector<uint8_t> bytes() const;
 
    private:
-      size_t m_leaf_idx;
+      uint64_t m_leaf_idx;
       secure_vector<uint8_t> m_randomness;
       XMSS_WOTS_PublicKey::TreeSignature m_tree_sig;
    };
